+++ src/test/java/com/vonage/client/OrderedMap.java
@@ -20,26 +20,26 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 
-public class OrderedJsonMap extends LinkedHashMap<String, Object> {
+public class OrderedMap extends LinkedHashMap<String, Object> {
 
-    public OrderedJsonMap() {
+    public OrderedMap() {
         super();
     }
 
     @SafeVarargs
-    public OrderedJsonMap(Entry<String, ?>... entries) {
+    public OrderedMap(Entry<String, ?>... entries) {
         super(entries.length);
         for (var entry : entries) {
             put(entry.getKey(), entry.getValue());
         }
     }
 
-    public OrderedJsonMap add(String key, Object value) {
+    public OrderedMap add(String key, Object value) {
         put(key, value);
         return this;
     }
 
-    public OrderedJsonMap addAll(Map<String, ?> other) {
+    public OrderedMap addAll(Map<String, ?> other) {
         putAll(other);
         return this;
     }