+++ src/test/java/com/vonage/client/auth/camara/NetworkAuthClientTest.java
@@ -17,21 +17,25 @@ package com.vonage.client.auth.camara;
 
 import com.vonage.client.AbstractClientTest;
 import com.vonage.client.DynamicEndpoint;
+import com.vonage.client.OrderedMap;
+import static com.vonage.client.OrderedMap.entry;
 import com.vonage.client.RestEndpoint;
 import static com.vonage.client.TestUtils.TEST_BASE_URI;
 import static com.vonage.client.TestUtils.testJsonableBaseObject;
 import static com.vonage.client.auth.camara.FraudPreventionDetectionScope.CHECK_SIM_SWAP;
 import static com.vonage.client.auth.camara.FraudPreventionDetectionScope.RETRIEVE_SIM_SWAP_DATE;
 import com.vonage.client.common.HttpMethod;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.*;
+import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.function.Executable;
+import java.net.URI;
 import java.util.Map;
 import java.util.UUID;
 
 public class NetworkAuthClientTest extends AbstractClientTest<NetworkAuthClient> {
-    final String authReqId = "arid/"+UUID.randomUUID();
+    final URI redirectUrl = URI.create("http://example.org/redirect");
+    final String authReqId = "arid/"+UUID.randomUUID(),
+            msisdn = "447700900001", state = "MyApp_state123";
 
     public NetworkAuthClientTest() {
         client = new NetworkAuthClient(wrapper);
@@ -55,13 +59,13 @@ public class NetworkAuthClientTest extends AbstractClientTest<NetworkAuthClient>
 
     @SuppressWarnings("unchecked")
     @Test
-    public void testFraudBackendAuthMethod() throws Exception {
-        var fbam = new FraudBackendAuthMethod(client, "+34 91 12345678", RETRIEVE_SIM_SWAP_DATE);
+    public void testNetworkAuthMethod() throws Exception {
+        var fbam = new NetworkAuthMethod(client, new BackendAuthRequest("+34 91 12345678", RETRIEVE_SIM_SWAP_DATE));
         wrapper.getAuthCollection().add(fbam);
         var endpoint = DynamicEndpoint.<Void, String> builder(String.class)
                 .wrapper(wrapper).requestMethod(HttpMethod.POST)
-                .authMethod(FraudBackendAuthMethod.class)
-                .pathGetter((de, req) -> TEST_BASE_URI).build();
+                .authMethod(NetworkAuthMethod.class)
+                .pathGetter((_, _) -> TEST_BASE_URI).build();
 
         var expectedResponse = "Hello, GNP!";
         stubNetworkResponse(expectedResponse);
@@ -70,8 +74,9 @@ public class NetworkAuthClientTest extends AbstractClientTest<NetworkAuthClient>
     }
 
     @Test
-    public void testBackendAuth() throws Exception {
+    public void testSendAuthRequest() throws Exception {
         final int expiresIn = 120, interval = 3;
+        final String state = "Unique app id";
         String responseJson = STR."""
             {
                "auth_req_id": "\{authReqId}",
@@ -79,26 +84,33 @@ public class NetworkAuthClientTest extends AbstractClientTest<NetworkAuthClient>
                "interval": "\{interval}"
             }
         """, msisdn = "+49 151 1234567";
-        var scope = FraudPreventionDetectionScope.RETRIEVE_SIM_SWAP_DATE;
-        var request = new BackendAuthRequest(msisdn, scope);
 
-        stubResponse(200, responseJson);
-        var parsed = client.sendAuthRequest(request);
+        var backendScope = FraudPreventionDetectionScope.RETRIEVE_SIM_SWAP_DATE;
+        var backend = new BackendAuthRequest(msisdn, backendScope);
+        var frontend = new FrontendAuthRequest(msisdn, redirectUrl, state);
 
-        testJsonableBaseObject(parsed);
-        assertEquals(authReqId, parsed.getAuthReqId());
-        assertEquals(expiresIn, parsed.getExpiresIn());
-        assertEquals(interval, parsed.getInterval());
+        for (var request : new AuthRequest[]{backend, frontend}) {
+            stubResponse(200, responseJson);
+            var parsed = client.sendAuthRequest(request);
+            testJsonableBaseObject(parsed);
+            assertEquals(authReqId, parsed.getAuthReqId());
+            assertEquals(expiresIn, parsed.getExpiresIn());
+            assertEquals(interval, parsed.getInterval());
+        }
 
         assertThrows(NullPointerException.class, () -> new BackendAuthRequest(msisdn, null));
-        assertThrows(NullPointerException.class, () -> new BackendAuthRequest(null, scope));
-        assertThrows(IllegalArgumentException.class, () -> new BackendAuthRequest("foo", scope));
+        assertThrows(NullPointerException.class, () -> new BackendAuthRequest(null, backendScope));
+        assertThrows(IllegalArgumentException.class, () -> new BackendAuthRequest("foo", backendScope));
+
+        assertThrows(NullPointerException.class, () -> new FrontendAuthRequest(null, redirectUrl, state));
+        assertThrows(NullPointerException.class, () -> new FrontendAuthRequest(msisdn, null, state));
+        assertNull(new FrontendAuthRequest(msisdn, redirectUrl, null).makeParams().get("state"));
 
         stubResponseAndAssertThrows(200, responseJson,
-                () -> client.sendAuthRequest(null), NullPointerException.class
+                () -> client.sendAuthRequest(null), IllegalArgumentException.class
         );
 
-        assert403ResponseException(() -> client.sendAuthRequest(request));
+        assert403ResponseException(() -> client.sendAuthRequest(backend));
     }
 
     @Test
@@ -131,11 +143,10 @@ public class NetworkAuthClientTest extends AbstractClientTest<NetworkAuthClient>
 
     @Test
     public void testBackendAuthEndpoint() throws Exception {
-        new NetworkAuthEndpointTestSpec<BackendAuthRequest, BackendAuthResponse>() {
-            final String msisdn = "447700900000";
+        new NetworkAuthEndpointTestSpec<BackendAuthRequest, AuthResponse>() {
 
             @Override
-            protected RestEndpoint<BackendAuthRequest, BackendAuthResponse> endpoint() {
+            protected RestEndpoint<BackendAuthRequest, AuthResponse> endpoint() {
                 return client.backendAuth;
             }
 
@@ -150,11 +161,60 @@ public class NetworkAuthClientTest extends AbstractClientTest<NetworkAuthClient>
             }
 
             @Override
-            protected Map<String, String> sampleQueryParams() {
-                return Map.of(
-                        "login_hint", "tel:" + msisdn,
-                        "scope", "dpv:FraudPreventionAndDetection#check-sim-swap"
+            protected Map<String, ?> sampleQueryParams() {
+                return new OrderedMap(
+                        entry("login_hint", "tel:+" + msisdn),
+                        entry("scope", "dpv:FraudPreventionAndDetection#check-sim-swap")
                 );
+            }
+
+            @Override
+            protected HttpMethod expectedHttpMethod() {
+                return HttpMethod.POST;
+            }
+        }
+        .runTests();
+    }
+
+    @Test
+    public void testFrontendAuthEndpoint() throws Exception {
+        new NetworkAuthEndpointTestSpec<FrontendAuthRequest, AuthResponse>() {
+
+            @Override
+            protected RestEndpoint<FrontendAuthRequest, AuthResponse> endpoint() {
+                return client.frontendAuth;
+            }
+
+            @Override
+            protected String customBaseUri() {
+                // This is hardcoded
+                return expectedDefaultBaseUri();
+            }
+
+            @Override
+            protected String expectedEndpointUri(FrontendAuthRequest request) {
+                return "/oauth2/auth";
+            }
+
+            @Override
+            protected FrontendAuthRequest sampleRequest() {
+                return new FrontendAuthRequest(msisdn, redirectUrl, state);
+            }
+
+            @Override
+            protected Map<String, ?> sampleQueryParams() {
+                return new OrderedMap(
+                        entry("login_hint", "tel:+" + msisdn),
+                        entry("scope", "dpv:FraudPreventionAndDetection#number-verification-verify-read"),
+                        entry("redirect_uri", redirectUrl.toString()),
+                        entry("state", state),
+                        entry("response_type", "code")
+                );
+            }
+
+            @Override
+            protected HttpMethod expectedHttpMethod() {
+                return HttpMethod.GET;
             }
         }
         .runTests();
@@ -180,11 +240,16 @@ public class NetworkAuthClientTest extends AbstractClientTest<NetworkAuthClient>
             }
 
             @Override
-            protected Map<String, String> sampleQueryParams() {
-                return Map.of(
-                        "grant_type", "urn:openid:params:grant-type:ciba",
-                        "auth_req_id", authReqId
+            protected Map<String, ?> sampleQueryParams() {
+                return new OrderedMap(
+                        entry("grant_type", "urn:openid:params:grant-type:ciba"),
+                        entry("auth_req_id", authReqId)
                 );
+            }
+
+            @Override
+            protected HttpMethod expectedHttpMethod() {
+                return HttpMethod.POST;
             }
         }
         .runTests();